Refactor the authentication system, using Redux
