+++ frontEnd/package.json
@@ -13,8 +13,10 @@
     "@react-native-community/masked-view": "0.1.10",
     "@react-navigation/native": "^5.9.4",
     "@react-navigation/stack": "^5.14.4",
+    "@reduxjs/toolkit": "^1.5.1",
     "@types/axios": "^0.14.0",
     "@types/react": "^17.0.3",
+    "@types/react-redux": "^7.1.16",
     "axios": "^0.21.1",
     "callstack": "^1.0.1",
     "expo": "~41.0.0",
@@ -32,6 +34,8 @@
     "react-native-safe-area-context": "3.2.0",
     "react-native-screens": "~3.0.0",
     "react-native-web": "~0.13.12",
+    "react-redux": "^7.2.3",
+    "redux": "^4.0.5",
     "use-debounce": "^6.0.1",
     "yup": "^0.32.9"
   },
