+++ frontEnd/src/components/Root/ActionType.tsx
@@ -1,0 +1,9 @@
+enum ActionType {
+    SIGN_IN = 0,
+    SIGN_UP = 1,
+    SIGN_OUT = 2,
+    
+    RESTORE_TOKEN = 3
+}
+
+export default ActionType
